student_module/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
pymysql==1.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
openai>=1.0.0
google-generativeai>=0.8.0
google-genai>=1.0.0
